#987 REFACTOR push colored diff into SafeOutput
